requirements.txt
pymongo
pandas 
numpy 
dotenv
certifi
scikit-learn
pyyaml
ipykernel
matplotlib
seaborn
statsmodels
imbalanced-learn
mlflow
keras
tensorflow